refactor(indexer): unexport createBinlogEventsTable, move its test home (B2 review fold-in #3)
The /review-pr type-design finding: CreateBinlogEventsTable was exported SOLELY
so a cmd integration test could call it — contradicting the #435 precedent
(buildFKCascadeQuery was unexported + its test relocated). Caught pre-merge, so
the unnecessary export never lands on main.

- internal/indexer/schema.go: CreateBinlogEventsTable → createBinlogEventsTable
  (private; same-package CreateIndexTables calls it fine without the export).
- TestCreateBinlogEventsTable moved cmd/bintrail/cmd_integration_test.go →
  new internal/indexer/schema_integration_test.go (package indexer,
  //go:build integration), calling the private name. testutil does NOT import
  indexer, so no cycle.
- The 3 cmd partition tests (TestLoadPartitionStats/ListPartitions/DropPartitions)
  only needed a partitioned binlog_events FIXTURE — switched from the now-private
  createBinlogEventsTable to the already-exported
  indexer.CreateIndexTables(context.Background(), db, N, false, nil). Same N+1
  partitions; t.Fatalf strings updated.

indexer's public provisioning surface is now exactly CreateIndexTables +
EnsureDatabase (what production/the supervisor actually call). Verified: build/
vet clean (unit + integration); full unit suite green; the moved
TestCreateBinlogEventsTable passes in internal/indexer + the 3 rewired cmd
fixture tests pass against MySQL.
